az-web-scrapper/requirements.txt
aiohttp==3.8.3
aiosignal==1.3.1
async-timeout==4.0.2
attrs==22.2.0
Automat==22.10.0
azure-functions==1.12.0
azure-functions-durable==1.2.1
bcrypt==4.0.1
certifi==2022.12.7
cffi==1.15.1
charset-normalizer==2.1.1
constantly==15.1.0
cryptography==38.0.4
cssselect==1.2.0
filelock==3.8.2
frozenlist==1.3.3
furl==2.1.3
hyperlink==21.0.0
idna==3.4
incremental==22.10.0
itemadapter==0.7.0
itemloaders==1.0.6
jmespath==1.0.1
lxml==4.9.2
multidict==6.0.3
orderedmultidict==1.0.1
packaging==22.0
paramiko==2.12.0
parsel==1.7.0
Protego==0.2.1
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.21
PyDispatcher==2.0.6
PyNaCl==1.5.0
pyOpenSSL==22.1.0
pysftp==0.2.9
python-dateutil==2.8.2
queuelib==1.6.2
requests==2.28.1
requests-file==1.5.1
Scrapy==2.7.1
service-identity==21.1.0
six==1.16.0
tldextract==3.4.0
Twisted==22.10.0
typing_extensions==4.4.0
urllib3==1.26.13
w3lib==2.1.1
yarl==1.8.2
zope.interface==5.5.2
